+++ models/v4_cluster_details_response_workers_items_memory.go
@@ -16,7 +16,7 @@ import (
 type V4ClusterDetailsResponseWorkersItemsMemory struct {
 
 	// RAM size in GB. Can be an integer or float.
-	SizeGb float64 `json:"size_gb,omitempty"`
+	SizeGb float32 `json:"size_gb,omitempty"`
 }
 
 // Validate validates this v4 cluster details response workers items memory
